Create param slice
