+++ src/inmanta/server/__init__.py
@@ -29,3 +29,4 @@ SLICE_PROJECT_ENV = "core.project"
 SLICE_FILE = "core.slice"
 SLICE_CODE = "core.code"
 SLICE_METRICS = "core.metrics"
+SLICE_PARAM = "core.parameters"
